adding check security on call API

+++ src/pages/archive/archive.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { NavController } from 'ionic-angular';
 import { CommonService } from '../../providers/common-service';
 import { MangaService } from '../../providers/manga-service';
+import { SecurityService } from '../../providers/security-service';
+import { LoginPage } from '../login/index';
 import { File } from '@ionic-native/file';
 import { FileTransfer, FileTransferObject } from '@ionic-native/file-transfer';
 import { LocalNotifications } from '@ionic-native/local-notifications';
@@ -10,10 +12,11 @@ import { Observable } from 'rxjs/Rx';
 @Component({
     selector: 'page-archive',
     templateUrl: 'archive.html',
-    providers: [CommonService, MangaService]
+    providers: [CommonService, MangaService, SecurityService]
 })
 export class ArchivePage {
 
+    private currentArchiveDownload:any = false;
     private archives:any = [];
     private fileTransfer:FileTransferObject = null;
     private subscription:any = 0;
@@ -23,17 +26,21 @@ export class ArchivePage {
 
     constructor(public navCtrl: NavController, public commonService: CommonService,
                 public mangaService: MangaService, private transfer: FileTransfer,
-                private file: File, private localNotifications: LocalNotifications) {
-        if(!this.fileTransfer) {
-            this.fileTransfer = this.transfer.create();
-        }
+                private file: File, private localNotifications: LocalNotifications,
+                private securityService: SecurityService) {
     }
 
     ionViewDidLoad () {
-        this.commonService.loadingShow('Please wait...');
-        this.showArchives();
-        this.subscription = Observable.interval(1000).subscribe(x => {
-            this.loadProgress();
+        this.securityService.checkAuth().then(auth => {
+            if (!auth) {
+                this.navCtrl.setRoot(LoginPage);
+            } else {
+                this.commonService.loadingShow('Please wait...');
+                this.showArchives();
+                this.subscription = Observable.interval(1000).subscribe(x => {
+                    this.loadProgress();
+                });
+            }
         });
     }
 
@@ -49,13 +56,17 @@ export class ArchivePage {
     }
 
     remove(archive) {
-        this.commonService.loadingShow('Please wait...');
-        this.mangaService.removeDownload(archive.id).then(remove => {
-            if (!remove) {
-                this.commonService.toastShow('Impossible de supprimer le téléchargement');
-            }
-            this.showArchives();
-        });
+        if (this.currentArchiveDownload) {
+            this.commonService.toastShow("Impossible de supprimer l'archive pendant son téléchargement");
+        } else {
+            this.commonService.loadingShow('Please wait...');
+            this.mangaService.removeDownload(archive.id).then(remove => {
+                if (!remove) {
+                    this.commonService.toastShow("Impossible de supprimer l'archive");
+                }
+                this.showArchives();
+            });
+        }
     }
 
     download(archive) {
@@ -68,18 +79,26 @@ export class ArchivePage {
                 let filename = data['name'];
                 this.filesize = data['size'];
                 this.mangaService.getUrlArchiveDownload(archive.id).then(url => {
-                    this.inProgress = true;
-                    this.fileTransfer.download(url.toString(), this.file.externalRootDirectory + '/Download/' + filename).then((entry) => {
-                        this.localNotifications.schedule({
-                            id: 1,
-                            text: 'Le fichier téléchargé a été déposé sous '+ entry.toURL(),
-                            sound: null
+                    if(this.fileTransfer==null) {
+                        this.fileTransfer = this.transfer.create();
+                        this.inProgress = true;
+                        this.currentArchiveDownload = true;
+                        this.fileTransfer.download(url.toString(), this.file.externalRootDirectory + '/Download/' + filename).then((entry) => {
+                            this.localNotifications.schedule({
+                                id: 1,
+                                text: 'Le fichier téléchargé a été déposé sous '+ entry.toURL(),
+                                sound: null
+                            });
+                            this.filesize = 0;
+                            this.inProgress = false;
+                            this.fileTransfer = null;
+                            this.currentArchiveDownload = false;
+                        }, (error) => {
+                            this.commonService.toastShow('Erreur : impossible de télécharger le fichier');
                         });
-                        this.filesize = 0;
-                        this.inProgress = false;
-                    }, (error) => {
-                        this.commonService.toastShow('Erreur : impossible de télécharger le fichier');
-                    });
+                    } else {
+                        this.commonService.toastShow('Veuillez patienter, un téléchargement est déjà en cours.');
+                    }
                 });
             }
             this.commonService.loadingHide();
